Add bootstrapping to incentive service; adapt spring configurations.
