+++ services/bootstrap/src/main/resources/application-dev.properties
@@ -1,4 +1,8 @@
-basket-service.provider-secret=xMcOZ4zPSeR5FeZ6lFtopdID53tpBNSX2TTfie9WUnwmFiBaYfFdo6RoX4mIcKxYHNhyoH3kcusYHecjDijfthS9Gv9APrN9Y14kqS7iSdSpo733gVNaEbFCJkABGQKh
 basket-service.url=http://localhost:8010
+basket-service.provider-secret=xMcOZ4zPSeR5FeZ6lFtopdID53tpBNSX2TTfie9WUnwmFiBaYfFdo6RoX4mIcKxYHNhyoH3kcusYHecjDijfthS9Gv9APrN9Y14kqS7iSdSpo733gVNaEbFCJkABGQKh
+
+incentive-service.url=http://localhost:8012
+incentive-service.provider-secret=hRIzOnQ6JL9LqYnVQoC4qdD2lADhtEbXSbdUfHspOGPKnECCbTBVrvk5pZAp7dJCuTUs3zqMnM4N56sKZnHDbSSEiGXX84g0uakStU6ofKWumCGyUg0PzVu75CeziWfY
+
 logging.level.org.springframework.web=DEBUG
 logging.level.org.hibernate=ERROR
