+++ services/bootstrap/src/main/resources/application-production.properties
@@ -1,2 +1,5 @@
-basket-service.provider-secret=
 basket-service.url=http://basketservice:8010/basket/
+basket-service.provider-secret=
+
+incentive-service.url=http://promotionservice:8012/promotion/
+incentive-service.provider-secret=
