+++ services/client/src/main/java/org/cryptimeleon/incentive/client/AliveEndpoint.java
@@ -1,0 +1,7 @@
+package org.cryptimeleon.incentive.client;
+
+import reactor.core.publisher.Mono;
+
+public interface AliveEndpoint {
+    Mono<String> sendAliveRequest();
+}
