+++ services/client/src/main/java/org/cryptimeleon/incentive/client/InfoClient.java
@@ -7,7 +7,7 @@ import reactor.core.publisher.Mono;
  * Client calls for info service.
  * Can be used for testing and prototyping.
  */
-public class InfoClient {
+public class InfoClient implements AliveEndpoint {
 
     /**
      * Webclient configured with the url of the info service
