+++ services/client/src/test/java/org/cryptimeleon/incentive/client/integrationtest/FullWorkflowTest.java
@@ -12,11 +12,11 @@ import org.cryptimeleon.incentive.crypto.model.IncentivePublicParameters;
 import org.cryptimeleon.incentive.crypto.model.keys.provider.ProviderPublicKey;
 import org.cryptimeleon.incentive.crypto.model.messages.JoinResponse;
 import org.cryptimeleon.incentive.promotion.Promotion;
-import org.cryptimeleon.incentive.promotion.sideeffect.RewardSideEffect;
 import org.cryptimeleon.incentive.promotion.hazel.HazelPromotion;
 import org.cryptimeleon.incentive.promotion.hazel.HazelTokenUpdate;
 import org.cryptimeleon.incentive.promotion.model.Basket;
 import org.cryptimeleon.incentive.promotion.model.BasketItem;
+import org.cryptimeleon.incentive.promotion.sideeffect.RewardSideEffect;
 import org.cryptimeleon.math.serialization.converter.JSONConverter;
 import org.cryptimeleon.math.structures.rings.RingElement;
 import org.junit.jupiter.api.Assertions;
@@ -61,7 +61,10 @@ public class FullWorkflowTest extends IncentiveSystemIntegrationTest {
                 publicParameters.getSpsEq(),
                 publicParameters.getBg().getG1());
         var incentiveSystem = new IncentiveSystem(publicParameters);
-        assertTrue(incentiveClient.addPromotions(List.of(testPromotion)).block(Duration.ofSeconds(1)).getStatusCode().is2xxSuccessful());
+        assertTrue(incentiveClient.addPromotions(List.of(testPromotion), sharedSecret)
+                .block(Duration.ofSeconds(1))
+                .getStatusCode()
+                .is2xxSuccessful());
 
         var userKeyPair = incentiveSystem.generateUserKeys();
