+++ services/client/src/test/resources/application.properties
@@ -2,6 +2,8 @@ dsprotection.url=http://localhost:8004
 basket.url=http://localhost:8010
 info.url=http://localhost:8000
 incentive.url=http://localhost:8012
+
+provider.shared-secret=CSOSopEXYJcGrc7NOi6M96kwCv1k2Tv0nL2lAce2jVBt5FaLjh3SeNslq6cX6nrl51iRv1PbAQUvnhOrpXrLgJZxHJrNRZE7vV1kWag73Gkm2IMG21i1858PAT4N5F11
 basket-service.redeem-secret=M6Dh805en61mYx52sIQsSu33VPJjjhiBhYlon12wbelvtpj8XkkezWftbfQBNCpzW0jBUcXws6qXE1nV78oGUE7sqslfe4XQ9Yz84iQor8vWmnvJqrXRad17wcCCybIa
 basket-service.pay-secret=2hc3VzsDDXrWCAPuYY7TQzrAwoVGSVh8z8FYWHb5pZlcKuhOAXnOGa0jmgRZHhNHjwTZkH3kJDJwAXF38ORkIp74N1yYQqfWZKZZppdFRrHSgHFU695cFeeWWBaPhapi
-provider.shared-secret=CSOSopEXYJcGrc7NOi6M96kwCv1k2Tv0nL2lAce2jVBt5FaLjh3SeNslq6cX6nrl51iRv1PbAQUvnhOrpXrLgJZxHJrNRZE7vV1kWag73Gkm2IMG21i1858PAT4N5F11
+incentive-service.provider-secret=hRIzOnQ6JL9LqYnVQoC4qdD2lADhtEbXSbdUfHspOGPKnECCbTBVrvk5pZAp7dJCuTUs3zqMnM4N56sKZnHDbSSEiGXX84g0uakStU6ofKWumCGyUg0PzVu75CeziWfY
