+++ services/inc/src/main/java/org/cryptimeleon/incentive/services/promotion/repository/PromotionRepository.java
@@ -1,15 +1,6 @@
 package org.cryptimeleon.incentive.services.promotion.repository;
 
 import org.cryptimeleon.incentive.promotion.Promotion;
-import org.cryptimeleon.incentive.promotion.hazel.HazelPromotion;
-import org.cryptimeleon.incentive.promotion.hazel.HazelTokenUpdate;
-import org.cryptimeleon.incentive.promotion.sideeffect.NoSideEffect;
-import org.cryptimeleon.incentive.promotion.sideeffect.RewardSideEffect;
-import org.cryptimeleon.incentive.promotion.streak.RangeProofStreakTokenUpdate;
-import org.cryptimeleon.incentive.promotion.streak.SpendStreakTokenUpdate;
-import org.cryptimeleon.incentive.promotion.streak.StandardStreakTokenUpdate;
-import org.cryptimeleon.incentive.promotion.streak.StreakPromotion;
-import org.cryptimeleon.incentive.promotion.vip.VipPromotion;
 import org.cryptimeleon.incentive.services.promotion.IncentiveServiceException;
 import org.springframework.stereotype.Repository;
 
@@ -17,7 +8,6 @@ import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 
 /**
  * This repository manages all promotions.
@@ -25,93 +15,7 @@ import java.util.UUID;
  */
 @Repository
 public class PromotionRepository {
-    private final List<Promotion> promotions = new ArrayList<>(
-            List.of(
-                    new HazelPromotion(
-                            HazelPromotion.generatePromotionParameters(),
-                            "Nutella Promotion",
-                            "Earn one point for every jar of Nutella purchased!",
-                            List.of(
-                                    new HazelTokenUpdate(
-                                            UUID.randomUUID(),
-                                            "Get a free hazelnut spread for 4 points!",
-                                            new RewardSideEffect("0580082614202"),
-                                            4
-                                    ),
-                                    new HazelTokenUpdate(
-                                            UUID.randomUUID(),
-                                            "Get a free large hazelnut spread for 9 points!",
-                                            new RewardSideEffect("4499722672546"),
-                                            9
-                                    )
-                            ),
-                            "hazel"
-                    ),
-                    new HazelPromotion(
-                            HazelPromotion.generatePromotionParameters(),
-                            "General Promotion",
-                            "Earn one point for every item you buy!",
-                            List.of(
-                                    new HazelTokenUpdate(
-                                            UUID.randomUUID(),
-                                            "Get a free Teddy for 500 points!",
-                                            new RewardSideEffect("4460463579054"),
-                                            500),
-                                    new HazelTokenUpdate(
-                                            UUID.randomUUID(),
-                                            "Get a free Pan for 1000 points!",
-                                            new RewardSideEffect("0916751964193"),
-                                            1000)
-                            ),
-                            ""
-                    ),
-                    new VipPromotion(
-                            VipPromotion.generatePromotionParameters(),
-                            "VIP Promotion",
-                            "You can reach the VIP status BRONZE, SILVER and Gold by collecting points for every purchase.",
-                            100_00, // costs are in cent, hence the _
-                            200_00,
-                            500_00,
-                            new RewardSideEffect("8445463753160"),
-                            new RewardSideEffect("0789590748887"),
-                            new RewardSideEffect("1393421332370")
-                    ),
-                    new StreakPromotion(
-                            StreakPromotion.generatePromotionParameters(),
-                            "Streak Promotion",
-                            "Maintain a streak by shopping regularly. You lose your streak if you do not visit our store for 7 days in a row!",
-                            List.of(
-                                    new StandardStreakTokenUpdate(
-                                            UUID.randomUUID(),
-                                            "Increase or reset your streak",
-                                            new NoSideEffect(), // This is just the default update operation
-                                            7
-                                    ),
-                                    new RangeProofStreakTokenUpdate(
-                                            UUID.randomUUID(),
-                                            "You get a free coffee if you're streak is at least 5.",
-                                            new RewardSideEffect("2413860782644"),
-                                            7,
-                                            5
-                                    ),
-                                    new SpendStreakTokenUpdate(
-                                            UUID.randomUUID(),
-                                            "Get a free manicure set in exchange for a streak of 10.",
-                                            new RewardSideEffect("0750769787791"),
-                                            7,
-                                            10
-                                    ),
-                                    new SpendStreakTokenUpdate(
-                                            UUID.randomUUID(),
-                                            "Get a free knife set in exchange for a streak of 20.",
-                                            new RewardSideEffect("0182420525002"),
-                                            7,
-                                            20
-                                    )
-                            ),
-                            7
-                    )
-            ));
+    private final List<Promotion> promotions = new ArrayList<>();
 
     public List<Promotion> getPromotions() {
         return promotions;
